fix: align replay, player, and talent table types with actual schema
Fixed remaining type mismatches in replay-related and talent tables to match
the actual database schema and ensure type consistency across joins.

Replay Table Updates:
- Changed replay.replayID from unsignedBigInteger to unsignedInteger to match
  actual production schema
- Changed replay.game_type from unsignedTinyInteger to tinyInteger to match
  actual production schema

Player Table Updates:
- Changed player.replayID from unsignedBigInteger to unsignedInteger to match
  replay.replayID
- Changed player.hero from unsignedTinyInteger to tinyInteger to match
  heroes.id
- Changed player_table_id from id() (bigint) to increments() (integer) for
  consistency

Talent Tables Updates:
- Changed talent_combinations.hero from integer to tinyInteger to match
  heroes.id
- Changed global_hero_talents_id from id() (bigint) to increments() (integer)
  for consistency

These changes resolve type mismatches in:
- GlobalTalentBuilderController (replay/player/talent joins)
- Ensures talents.replayID matches replay.replayID

Benefits:
- Eliminates implicit type conversions in joins between replay, player, and
  talents tables
- Ensures hero columns consistently use tinyInteger across all tables
- Aligns migrations with actual production database schema
- Maintains auto-increment primary key consistency (integer instead of bigint)

+++ database/migrations/2025_08_08_184432_create_talent_combinations_table_heroesprofile.php
@@ -15,7 +15,7 @@ class CreateTalentCombinationsTableHeroesprofile extends Migration
     {
         Schema::create('talent_combinations', function (Blueprint $table) {
             $table->increments('talent_combination_id');
-            $table->integer('hero')->nullable();
+            $table->tinyInteger('hero')->nullable();
             $table->integer('level_one')->nullable();
             $table->integer('level_four')->nullable();
             $table->integer('level_seven')->nullable();
